requirements.txt
python-telegram-bot
requests
lxml
python-dotenv
urllib3
pylint
pytest
pytest-asyncio
PyYAML
